+++ variables-kv.tf
@@ -52,16 +52,12 @@ variable "keyvault_reader_objects_ids" {
 variable "keyvault_network_acls" {
   description = "Object with attributes: `bypass`, `default_action`, `ip_rules`, `virtual_network_subnet_ids`. See https://www.terraform.io/docs/providers/azurerm/r/key_vault.html#bypass for more informations."
   type = object({
-    bypass                     = string,
-    default_action             = string,
+    bypass                     = optional(string, "None"),
+    default_action             = optional(string, "Deny"),
     ip_rules                   = optional(list(string)),
-    virtual_network_subnet_ids = optional(list(string))
+    virtual_network_subnet_ids = optional(list(string)),
   })
-  default = {
-    default_action = "Deny"
-    bypass         = "None"
-  }
-
+  default = {}
 }
 
 # Keyvault Diagnotics